feat: 카운트다운 유틸 기반 게으른 초기화 및 suppressHydrationWarning 처리
